requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
db-sqlite3
jwt
PyJWT